+++ SobrassadaEngine/FileSystem/Meta/MetaFile.cpp
