+++ SobrassadaEngine/FileSystem/Meta/MetaTexture.cpp
