+++ SobrassadaEngine/FileSystem/Meta/MetaTexture.h
